Create Technical Ratings Pro - Pump Wave.pine
https://www.tradingview.com/script/A9tEtrNW-Technical-Ratings-Pro-Pump-Wave/
Jun 4, 2021
This script uses the built in Technical Ratings indicator but interprets the data visually. It plots the results for "total", "MA" and "other" as pump waves. It uses MA to plot a trend line (can be turned off in settings) . Candles are colored to the rating strength and a percentage number was added to the results. For more informations on the Technical Ratings indicator please refer to official documentation.
Feb 6
Release Notes: Minor update includes options to turn off each individual pump wave.