Update Center Of Gravity Oscillator.pine
Jun 28, 2021
The COG Oscillator (center of gravity ) is an indicator based on statistics and the Fibonacci golden ratio. It uses ALMA as a trigger and LSMA as "zero line". The trigger is set tight by default but can be tweaked by adjusting the window size and sigma in settings. This is a great indicator for setting up trades and spotting reversals. There are 2 main strategies that come with this indicator:

snapshot
Strategy 1: Long positions are entered when current low point is higher than previous low. Short positions are entered as current high is lower than previous high. (Shown in image above)

snapshot
Strategy 2 : If market is bullish long trades are entered as COG line crosses over red LSMA line. Traders have the option of scalping the first crossover or even scaling out of trade to close on second exit. This works the opposite for shorts when market is bearish .

snapshot
Above shows different configurations of the indicator. Top shows length of 50, Middle has length of 21 and bottom is default 9.
Jun 29, 2021
Release Notes: Added some extra features that can be turned on in settings:
- Fibonacci Levels
- RMA smoothing
- High/Low levels
- Price line (helps identify price breaking previous high/low)
snapshot
Jul 3, 2021
Release Notes: Added 2 new themes : Fibonacci and Fibonacci 2
Aug 21, 2021
Release Notes: Added "clean" theme. This theme is the same settings I use, cog length: 9, smoothing on with smoothing length: 9. It also uses SMA as the trigger.Sometimes I also use cog length: 50 to find more significant pivots.